requirements.txt
tweepy==4.14.0
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
python-multipart==0.0.9
jinja2==3.1.3
